Overview.html

<!DOCTYPE html>
<html lang="en"> 
<head> 
  <title>Testing How To</title> 
  <meta charset="utf-8">
  <meta name="copyright" 
   content="Copyright &#169; 2012  W3C. This work is distributed under the W3C® Software License." >
  <meta name="rights" content="see http://www.w3.org/Consortium/Legal/2002/copyright-software-20021231">
  <link rel="stylesheet" type="text/css" media="screen, projection, print" 
   href="http://www.w3.org/Talks/Tools/Slidy2/styles/slidy.css"> 
  <script src="http://www.w3.org/Talks/Tools/Slidy2/scripts/slidy.js"></script> 
  <style> 
   @media screen and (min-width: 700px) {
     div.slide {
       margin: 5%;
     }
   }
  </style> 
</head>
<body>
<div class="slide cover">
 <h1 style='margin-top: 15%'>Testing How To</h1> 

 <p><a href='https://github.com/w3c/testing-how-to'>https://github.com/w3c/testing-how-to</a></p>
</div> 

<div class="slide">
 <h1>Content</h1>
 <ol>
   <li>What do you test in a specification?</li>
   <li>How to write a test?</li>
   <li>W3C Test repositories</li>
   <li>How to add a test suite into the W3C framework?</li>
   <li>Licenses For W3C Test Suites</li>
   <li>Strategies for testing a specification</li>
 </ol>
</div>

<div class="slide">
 <h1>What do you test in a specification?</h1>

  <p>[Content]</p>
</div>

<div class="slide">
 <h1>How to write a test?</h1>

 <ol type="lower-alpha">
  <li>Script Test</li>
  <li>Reftests</li>
  <li>Self-Describing Test</li>
  <li>Client-Server Test</li>
 </ol>
</div>

<div class='slide'>
<h1>Script Test</h1>
<p>When the result of your test could be determined by a script.</p>
<pre style='font-size: larger'>&lt;script
  src="<a href='http://w3c-test.org/resources/testharness.js'>/resources/testharness.js</a>">
&lt;/script>
&lt;script
  src="<a href='http://w3c-test.org/resources/testharnessreport.js'>/resources/testharnessreport.js</a>">
&lt;/script></pre>

</div>

<div class='slide'>
<h1>testharness.js</h1>

<pre>
&lt;p id="MyE">This test checks the
  &lt;span>firstElementChild&lt;/span> property.&lt;/p>
&lt;div <em style='color: green'>id='log'</em>>&lt;/div>
&lt;script>
<em style='color: green'>test</em>(
 function() {
  var myE = document.getElementById("MyE");
  <em style='color: green'>assert_true</em>(!!MyE.firstElementChild);
 },
 "firstElementChild must exist"
);
&lt;/script>
</pre>

<p>See <a href='http://w3c-test.org/webperf/tests/approved/navigation-timing/html5/test_timing_attributes_exist.html'>Demo</a></p>
</div>

<div class='slide'>
<h1>testharness.js: assertions</h1>
<p>
<code>assert_true</code>,
<code>assert_false</code>,
<code>assert_equals</code>,
<code>assert_not_equals</code>,
<code>assert_approx_equals</code>
</p>
<p>
<code>assert_in_array</code>,
<code>assert_array_equals</code>,
<code>assert_regexp_equals</code>
</p>
<p>
<code>assert_own_property</code>,
<code>assert_inherits</code>,
<code>assert_idl_attribute</code>,
<code>assert_readonly</code>,
<code>assert_throws</code>,
<code>assert_unreached</code>,
<code>assert_any</code>
</p>
</div>

<div class='slide'>
<h1>testharness.js: async test</h1>

<pre>
var t = <em style='color: green'>async_test</em>("firstElementChild must exist");

function run() {
  t.<em style='color: green'>step</em>(
    function () {
      assert_true(!!MyE.firstElementChild);
    }
  );
  t.<em style='color: green'>done</em>();
};
window.onload = run;
</pre>

<p>See <a href='http://w3c-test.org/webapps/WebMessaging/tests/submissions/Microsoft/event.data.htm'>Demo</a> @@find a better example...</p> 
</div>

<div class='slide'>
<h1>testharness.js: metadata</h1>

<p>In the &lt;head> element:
<pre>
&lt;link rel="help"
 href="http://www.w3.org/TR/domcore/#dom-element-firstelementchild"/>
&lt;meta name="assert"
 content="The firstElementChild property must exist"/>
&lt;link rel="author"
 title="John Doe" href="mailto:john@doe.com"/>
</pre>
</div>

<div class='slide'>
<h1>testharness.js: per-test metadata</h1>

<pre>&lt;script>
<em style='color: green'>test</em>(
 function() {
  var myE = document.getElementById("MyE");
  <em style='color: green'>assert_true</em>(!!MyE.firstElementChild);
 },
 "firstElementChild",
 {
  help: 'http://www.w3.org/TR/domcore/#dom-element-firstelementchild', 
  assert: ['The firstElementChild property must exist'],
  author: 'John Doe <john@doe.com>' }
);
&lt;/script></pre>
</div>

<div class='slide'>
<h1>testharness.js: further readings</h1>

<ul>
<li>Read the <a href='http://w3c-test.org/resources/testharness.js'>documentation in the source</a></li>
<li><a href='http://darobin.github.com/test-harness-tutorial/docs/using-testharness.html'>testharness.js tutorial</a></li>
<li><a href='http://wiki.csswg.org/test/scripttest'>Writing Script Tests</a></li>
</ul>
</div>

<div class='slide'>
<h1>Reftest</h1>

<p>When the result of your test is visual.</p>

<table id='cssFirstLetter'>
<tr>
<th>bdo-000.html</th>
<th>bdo-000-ref.html</th>
</tr>
<tr class='lines'>
<td style='width: 45%'>
<bdo dir="rtl">HEBREW</bdo>
</td>
<td style='width: 45%'>
<span dir="ltr">WERBEH</span>
</td>
</tr>
<tr>
<td>
<pre>
&lt;bdo dir="rtl">HEBREW&lt;/bdo>
</pre>
</td>
<td>
<pre>
&lt;span>WERBEH&lt;/span>
</pre>
</td>
</tr>
</table>
</div>

<div class='slide'>
<h1>reftest: metadata</h1>

<p>As for testharness.js, put it in the &lt;head> element:
<pre>
&lt;link rel="help"
 href="http://www.w3.org/TR/domcore/#dom-element-firstelementchild"/>
&lt;meta name="assert"
 content="The firstElementChild property must exist"/>
&lt;link rel="author"
 title="John Doe" href="mailto:john@doe.com"/>
</pre>
</div>

<div class='slide'>
<h1>Reftest: further readings</h1>

  <ul>
   <li><a href='http://wiki.csswg.org/test/reftest'>Writing Reftests</a></li>
  </ul>
</div>

<div class='slide'>
<h1>Manual test</h1>

<p>When everything else wouldn't work.</p>

<div>
 <h2>visibilitychange</h2>
 <p>Iconify this window and restore it.</p>
 <div id='visibility' style='width: 50px; height: 50px;background-color: red;margin: auto;'></div>

 <p>If the square is still red, please indicate that the test has failed.</p>

<script>
document.addEventListener("visibilitychange",
 function() {
   document.getElementById("visibility").style.backgroundColor = 'green';
 }, false);
</script>

<p>See also <a href='http://w3c-test.org/webperf/tests/submission/Microsoft/PageVisibility/test_minimize.htm'>Demo</a></p>

</div>

</div>

<div class="slide">
 <h1>Client-Server test</h1>

 <p>When you need server-side code as well</p>

 <p>So far, we have two examples of those:
 <ol>
  <li>XMLHttpRequest</li>
  <li>The WebSocket API</li>
 </ol>
</div>

<div class="slide">
 <h1>XHR test</h1>

 <p>Use <a href='https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources'>PHP</a> to simulate server responses:</p>

 <ol>
   <li><a href="https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources/status.php">status.php</a> for content, http code, and media type (see <a href='http://w3c-test.org/webapps/XMLHttpRequest/tests/submissions/Opera/resources/status.php?content=%3Cx%3E%3C/x%3E&type=application/xml'>example</a>)</li>
   <li><a href="https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources/delay.php">delay.php</a> to delay HTTP responses</li>
   <li>&hellip;</li>
</ol>
</div>

<div class="slide">
 <h1>The WebSocket API</h1>

 <p>[content]</p>

</div>

<div class='slide'>
  <h1>Audio and video materials</h1>

  <p>If you're testing audio or video content, we have content already:</p>

  <ol>
    <li><a href='http://media.w3.org/2010/05/'>2010/05</a>: Open
    project movies, time ticker sounds, timed videos, single frame
      videos, etc.</li>
    <li><a href='http://media.w3.org/2011/01/'>2011/01</a>: sounds, timed videos</li>
    <li><a href='http://media.w3.org/2011/09/audio'>2011/09</a>: multi audio tracks</li>
  </ol>
</div>


<div class="slide">
 <h1>W3C Test repositories</li>

 <p>[hg, github]</p>
</div>

<div class="slide">
 <h1>How to add a test suite into the W3C framework</li>

 <p>[w3c-test.org]</p>
</div>

<div class="slide">
 <h1>Licenses For W3C Test Suites</h1>

  <p>W3C makes test suites available under two distinct licenses:
<ol>
<li>a <a href="http://www.w3.org/Consortium/Legal/2008/04-testsuite-license.html">W3C Test Suite License</a> for
    an Authoritative W3C Test Suite or when claims of performance with
    respect to a specification are required.</li>
<li>a <a href="http://www.w3.org/Consortium/Legal/2008/03-bsd-license.html">3-clause
BSD License</a> for other applications.</li>
</ol>

</div>

<div class="slide">
 <h1>Contributions For W3C Test Suites</h1>

 <ol>
   <li>If you're a participant of the Working Group, you already
   agreed to the requirements, so no special action is required in
   order to contribute.</li>
   <li>otherwise,
use <a href="http://www.w3.org/2002/09/wbs/1/testgrants-200409/">Grant&#xA0;I</a>
or <a href="http://www.w3.org/2002/09/wbs/1/testgrants2-200409/">Grant&#xA0;II</a>
to agree with the requirements.</li>

</div>

<div class="slide">
 <h1>Strategies for testing a specification</h1>

 <p>[content]</p>

</div>

<div class="slide">
 <h1>For more information</h1>

 <p>Use <a href='http://lists.w3.org/Archives/Public/public-test-infra/'>public-test-infra@w3.org</a></p>

</div>

</body>
</html>
